Move a lot of helper methods from LEDStripUtils.kt to AnimatedStripUtils.kt to reflect what they're extending
